% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asCrm.R
\name{asCrm}
\alias{asCrm}
\title{Associate 'CRM' Object}
\usage{
asCrm(apiKey, action, from, to, definitionId)
}
\arguments{
\item{apiKey}{An 'API' key that can be generated from a 'HubSpot' instance.}

\item{action}{Valid values are "Add" & "Delete".}

\item{from}{The internal ID of the object to be associated from.}

\item{to}{The internal ID of the object to be associated to.}

\item{definitionId}{The ID that codifies what assciation to be made. See \href{https://legacydocs.hubspot.com/docs/methods/crm-associations/crm-associations-overview}{HubSpot Documentation}}
}
\description{
This function allows you to quickly create associations between objects on 'HubSpot' such as company, contact, and deal.
}
\examples{
result <- asCrm(
  apiKey = apiKey,
  from = vid,
  to = companyId,
  definitionId = 1)
}
