% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/companyListAll.R
\name{companyListAll}
\alias{companyListAll}
\title{Retrieve ALL 'HubSpot' Current Companies}
\usage{
companyListAll(accessToken, properties = c("name", "website"))
}
\arguments{
\item{accessToken}{An 'Access Token' can be generated from a 'HubSpot' instance through a private app.}

\item{properties}{A vector of character values corresponds to 'HubSpot' internal names of the properties. Use 'companyProperties()' to check the internal names of the company properties. Defaults to 'c("name", "website")'.}
}
\description{
This function allows you to quickly retrieve all companies from 'HubSpot' via 'API' connection and output them as a dataframe.
}
\examples{
companyListAll <- companyListAll(accessToken = "accessToken")
}
