% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/companyProperties.R
\name{companyProperties}
\alias{companyProperties}
\title{Retrieve 'HubSpot' Company Properties}
\usage{
companyProperties(apiKey = apiKey, maxOptions = 50)
}
\arguments{
\item{apiKey}{An API key that can be generated on a HubSpot instance.}

\item{maxOptions}{The maximum number of options of a given property to be included in the output dataframe. Defaults to 50.}
}
\description{
This function allows you to quickly retrieve company properties from 'HubSpot' via 'API' connection.
}
\examples{
companyProperties <- companyProperties(apiKey = "demo", maxOptions = 100)
}
