% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactList.R
\name{contactList}
\alias{contactList}
\title{Retrieve 'HubSpot' Current Contacts}
\usage{
contactList(
  apiKey = apiKey,
  properties = c("firstname", "lastname", "email", "company"),
  limit = "Default"
)
}
\arguments{
\item{apiKey}{An API key that can be generated on a HubSpot instance.}

\item{properties}{A vector of character values corresponds to 'HubSpot' interval names of the properties. Use 'contactProperties()' to check the internal names of the company properties. Defaults to 'c("firstname", "lastname", "email", "company")'.}

\item{limit}{An integer to limit the number of companies to be retrieved. Defaults to no limit.}
}
\description{
This function allows you to quickly retrieve contact list from 'HubSpot' via 'API' connection.
}
\examples{
contactList <- contactList(apiKey = "demo", properties = c("firstname", "lastname", "email", "company"), limit = 2)
}
