% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactProperties.R
\name{contactProperties}
\alias{contactProperties}
\title{Retrieve 'HubSpot' Contact Properties}
\usage{
contactProperties(apiKey = apiKey, maxOptions = 50)
}
\arguments{
\item{apiKey}{An API key that can be generated on a HubSpot instance.}

\item{maxOptions}{The maximum number of options of a given property to be included in the output dataframe. Defaults to 50.}
}
\description{
This function allows you to quickly retrieve contact properties from 'HubSpot' via 'API' connection.
}
\examples{
contactProperties <- contactProperties(apiKey = "demo", maxOptions = 100)
}
