% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createContacts.R
\name{createContacts}
\alias{createContacts}
\title{Create or Add New Contacts on 'HubSpot'}
\usage{
createContacts(accessToken, data)
}
\arguments{
\item{accessToken}{An 'Access Token' can be generated from a 'HubSpot' instance through a private app.}

\item{data}{A dataframe that contains the contact's email and the properties associated with the new contact. The field names in the data must match (case sensitive) the property's internal name on HubSpot', use 'contactProperties()'.}
}
\description{
This function allows you to create or add new contacts from a dataframe to 'HubSpot' via 'API' connection.
}
\examples{
output <- createContacts(
  accessToken = accessToken,
  data = data)
}
