% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDeals.R
\name{createDeals}
\alias{createDeals}
\title{Create or Add New Deals on 'HubSpot'}
\usage{
createDeals(apiKey, data)
}
\arguments{
\item{apiKey}{An 'API' key that can be generated from a 'HubSpot' instance.}

\item{data}{A dataframe that contains the deal's name and the properties associated with the new deal. The field names in the data must match (case sensitive) the property's internal name on HubSpot', use 'dealProperties()'.}
}
\description{
This function allows you to create or add new deals from a dataframe to 'HubSpot' via 'API' connection.
}
\examples{
output <- createDeals(
  apiKey = apiKey,
  data = data)
}
