% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPdp.R
\name{createPdp}
\alias{createPdp}
\title{Create 'PDP' Policies}
\usage{
createPdp(clientID, secret, data)
}
\arguments{
\item{clientID}{The ID that can be generated through the 'DOMO' developer page.}

\item{secret}{A secret that can be generated through the 'DOMO' developer page.}

\item{data}{The dataframe to update the 'PDP' list on a dataset, should include "Dataset ID", "Policy Name", "Policy Column", "User ID", and "Policy Value".}
}
\description{
This function allows you to quickly create new 'PDP' policies on a certain 'DOMO' dataset via 'API' connection.
}
\examples{
createdPdp <- createPdp(
  clientID = "CLIENT ID",
  secret = "SECRET",
  data = "DATAFRAME")
}
