% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealProperties.R
\name{dealProperties}
\alias{dealProperties}
\title{Retrieve 'HubSpot' Deal Properties}
\usage{
dealProperties(apiKey, maxOptions = 50, readOnlyValue = FALSE)
}
\arguments{
\item{apiKey}{An 'API' key that can be generated from a 'HubSpot' instance.}

\item{maxOptions}{The maximum number of options of each property to be included in the output dataframe. Defaults to 50.}

\item{readOnlyValue}{A Boolean or Logical input to specify whether to include the readOnlyValue properties or not. Defaults to 'FALSE' which excludes read-only properties. 'TRUE' will include read-only properties.}
}
\description{
This function allows you to quickly retrieve Deal properties from 'HubSpot' via 'API' connection and output them as a dataframe.
}
\examples{
dealProperties <- dealProperties(
  apiKey = "demo",
  maxOptions = 100,
  readOnlyValue = TRUE)
}
