% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deletePdp.R
\name{deletePdp}
\alias{deletePdp}
\title{Delete 'PDP' Policies}
\usage{
deletePdp(clientID, secret, data)
}
\arguments{
\item{clientID}{The ID that can be generated through the 'DOMO' developer page.}

\item{secret}{A secret that can be generated through the 'DOMO' developer page.}

\item{data}{The dataframe that includes the "Dataset ID" and the "Policy ID" of the policies.}
}
\description{
This function allows you to quickly delete 'PDP' policies on a certain 'DOMO' dataset via 'API' connection.
}
\examples{
deletedPdp <- deletePdp(
  clientID = "CLIENT ID",
  secret = "SECRET",
  data = "DATAFRAME")
}
