% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPdp.R
\name{extractPdp}
\alias{extractPdp}
\title{Extract 'PDP' List}
\usage{
extractPdp(clientID, secret, dsID)
}
\arguments{
\item{clientID}{The ID that can be generated through the 'DOMO' developer page.}

\item{secret}{A secret that can be generated through the 'DOMO' developer page.}

\item{dsID}{The ID of a DOMO dataset that can be found on the URL of the dataset.}
}
\description{
This function allows you to quickly retrieve 'PDP' list from a certain 'DOMO' dataset via 'API' connection and output them as a dataframe.
}
\examples{
extractedPdp <- extractPdp(
  clientID = "CLIENT ID",
  secret = "SECRET",
  dsID = "DATASET ID")
}
