% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listAllPdp.R
\name{listAllPdp}
\alias{listAllPdp}
\title{Create 'PDP' List}
\usage{
listAllPdp(data)
}
\arguments{
\item{data}{The dataframe to update the 'PDP' list on a dataset, should include "Policy Name", "Policy Column", "User ID", and "Policy Value". "Policy ID" is optional.}
}
\description{
This function allows you to quickly create a list of 'PDP' from a dataframe.
}
\examples{
listedAllPdp <- listAllPdp(data = "DATAFRAME")
}
