% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listOnePdp.R
\name{listOnePdp}
\alias{listOnePdp}
\title{Create A Single 'PDP'}
\usage{
listOnePdp(data)
}
\arguments{
\item{data}{The dataframe to update the 'PDP' list on a dataset, should include "Policy Name", "Policy Column", "User ID", and "Policy Value". "Policy ID" is optional.}
}
\description{
This function allows you to quickly create a single 'PDP' from a dataframe.
}
\examples{
listedOnePdp <- listOnePdp(data = "DATAFRAME")
}
