% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propertyList.R
\name{propertyList}
\alias{propertyList}
\title{Retrieve 'HubSpot' Properties of A Certain Object}
\usage{
propertyList(accessToken, object_type, archived = FALSE)
}
\arguments{
\item{accessToken}{An 'Access Token' can be generated from a 'HubSpot' instance through a private app.}

\item{object_type}{A string input to specify what object the properties need to be pulled from such as "companies", "contacts", or "deals".}

\item{archived}{A Boolean or Logical input to specify whether pulling the archived properties or not. Defaults to 'FALSE' which excludes archived properties. 'TRUE' will include archived properties.}
}
\description{
This function allows you to quickly retrieve company properties from 'HubSpot' via 'API' connection and output them as a dataframe.
}
\examples{
propertyList <- propertyList(
  accessToken = "accessToken",
  object_type = "companies",
  archived = FALSE)
}
