% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorderPdp.R
\name{reorderPdp}
\alias{reorderPdp}
\title{Reorder 'PDP' List}
\usage{
reorderPdp(data)
}
\arguments{
\item{data}{The dataframe to reorder, should include "Policy Name", "Policy Column", "User ID", and "Policy Value".}
}
\description{
This function allows you to prepare your dataframe to be compared to the current 'PDP' list on DOMO by reordering the user id and policy value.
}
\examples{
reorderedPdp <- reorderPdp(data = "DATAFRAME")
}
