% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unixToDate.R
\name{unixToDate}
\alias{unixToDate}
\title{Convert 'UNIX' to Date}
\usage{
unixToDate(unix, tz = "GMT", type = "Second")
}
\arguments{
\item{unix}{The 'UNIX' values to be converted to date.}

\item{tz}{Stands for 'Time Zone' which defaults to 'GMT'. This input specifies the time zone of the 'UNIX' values.}

\item{type}{The second granularity of the 'UNIX' values. Valid inputs include: 'Millisecond', 'Microsecond', and 'Nanosecond'.}
}
\description{
This function allows you to conveniently convert 'UNIX' values to date values.
}
\examples{
output <- dateToUnix(
  unix = unix,
  tz = "GMT",
  type = "Millisecond")
}
