% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCompanies.R
\name{updateCompanies}
\alias{updateCompanies}
\title{Update Companies' Properties on 'HubSpot'}
\usage{
updateCompanies(accessToken, data)
}
\arguments{
\item{accessToken}{An 'Access Token' can be generated from a 'HubSpot' instance through a private app.}

\item{data}{A dataframe that contains the company's internal IDs and the properties associated with the companies. The field names in the data must match (case sensitive) the property's internal name on HubSpot', use 'companyProperties()'.}
}
\description{
This function allows you to update the properties of companies on 'HubSpot' via 'API' connection.
}
\examples{
output <- updateCompanies(
  accessToken = accessToken,
  data = data)
}
