% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateContacts.R
\name{uupdateContacts}
\alias{uupdateContacts}
\title{Update Contacts' Properties on 'HubSpot'}
\usage{
uupdateContacts(apiKey, data)
}
\arguments{
\item{apiKey}{An 'API' key that can be generated from a 'HubSpot' instance.}

\item{data}{A dataframe that contains the internal IDs of the contacts and the properties associated with the contacts. The field names in the data must match (case sensitive) the property's internal name on HubSpot', use 'contactProperties()'.}
}
\description{
This function allows you to update the properties of contacts from a dataframe on 'HubSpot' via 'API' connection.
}
\examples{
output <- updateContacts(
  apiKey = apiKey,
  data = data)
}
