% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/companyList.R
\name{companyList}
\alias{companyList}
\title{Retrieve 'HubSpot' Current Companies}
\usage{
companyList(
  apiKey = apiKey,
  properties = c("name", "website"),
  limit = "Default"
)
}
\arguments{
\item{apiKey}{An API key that can be generated on a HubSpot instance.}

\item{properties}{A vector of character values corresponds to 'HubSpot' interval names of the properties. Use 'companyProperties()' to check the internal names of the company properties. Defaults to 'c("name", "website")'.}

\item{limit}{An integer to limit the number of companies to be retrieved. Defaults to no limit.}
}
\description{
This function allows you to quickly retrieve company list from 'HubSpot' via 'API' connection.
}
\examples{
companyList <- companyList(apiKey = "demo", properties = c("name", "website", "is_public"), limit = 2)
}
