% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactList.R
\name{contactList}
\alias{contactList}
\title{Retrieve 'HubSpot' Current Contacts}
\usage{
contactList(
  accessToken,
  properties = c("firstname", "lastname", "email", "company"),
  limit = 10,
  after = "First Page"
)
}
\arguments{
\item{accessToken}{An 'Access Token' can be generated from a 'HubSpot' instance through a private app.}

\item{properties}{A vector of character values corresponds to 'HubSpot' interval names of the properties. Use 'contactProperties()' to check the internal names of the company properties. Defaults to 'c("firstname", "lastname", "email", "company")'.}

\item{limit}{An integer to limit the number of companies to be retrieved. Defaults to "100".}

\item{after}{'HubSpot' paginates the output of contact list into batches of 100. If your 'HubSpot' instance have more than 100 companies, you will need to include the offset code to get the next batch of 100 companies. Offset value of the page will be included in the result of this function, you can then use that to get the next page and so on.}
}
\description{
This function allows you to quickly retrieve contact list from 'HubSpot' via 'API' connection and output them as a dataframe.
}
\examples{
contactList <- contactList(
  accessToken = "accessToken",
  properties = c("firstname", "lastname", "email", "company"),
  limit = 10,
  after = "First Page")
}
