% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCompanies.R
\name{createCompanies}
\alias{createCompanies}
\title{Create or Add New Companies on 'HubSpot'}
\usage{
createCompanies(apiKey, data)
}
\arguments{
\item{apiKey}{An 'API' key that can be generated from a 'HubSpot' instance.}

\item{data}{A dataframe that contains the company's name and the properties associated with the new company. The field names in the data must match (case sensitive) the property's internal name on HubSpot', use 'companyProperties()'.}
}
\description{
This function allows you to create or add new companies from a dataframe to 'HubSpot' via 'API' connection.
}
\examples{
output <- createCompanies(
  apiKey = apiKey,
  data = data)
}
