% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealList.R
\name{dealList}
\alias{dealList}
\title{Retrieve 'HubSpot' Current Deals}
\usage{
dealList(
  apiKey = apiKey,
  properties = c("dealname", "dealstage", "amount", "closedate"),
  limit = "Default"
)
}
\arguments{
\item{apiKey}{An API key that can be generated on a HubSpot instance.}

\item{properties}{A vector of character values corresponds to 'HubSpot' interval names of the properties. Use 'dealProperties()' to check the internal names of the company properties. Defaults to 'c("dealname", "dealstage", "amount", "closedate")'.}

\item{limit}{An integer to limit the number of companies to be retrieved. Defaults to no limit.}
}
\description{
This function allows you to quickly retrieve deal list from 'HubSpot' via 'API' connection.
}
\examples{
dealList <- dealList(apiKey = "demo", properties = c("firstname", "lastname", "email", "company"), limit = 2)
}
