% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealProperties.R
\name{dealProperties}
\alias{dealProperties}
\title{Retrieve 'HubSpot' Deal Properties}
\usage{
dealProperties(apiKey = apiKey, maxOptions = 50)
}
\arguments{
\item{apiKey}{An API key that can be generated on a HubSpot instance.}

\item{maxOptions}{The maximum number of options of a given property to be included in the output dataframe. Defaults to 50.}
}
\description{
This function allows you to quickly retrieve Deal properties from 'HubSpot' via 'API' connection.
}
\examples{
dealProperties <- dealProperties(apiKey = "demo", maxOptions = 100)
}
